+++ cmd/serve/main.go
@@ -7,8 +7,8 @@ import (
 	"os"
 
 	"github.com/dhis2-sre/go-rate-limite/pgk/config"
-	"github.com/dhis2-sre/go-rate-limite/pgk/handler"
 	"github.com/dhis2-sre/go-rate-limite/pgk/proxy"
+	"github.com/dhis2-sre/go-rate-limite/pgk/rate"
 	"github.com/dhis2-sre/go-rate-limite/pgk/rule"
 )
 
@@ -19,16 +19,19 @@ func main() {
 	}
 }
 
+// TODO: for testability you can pass in the io.Writer for logging, and any
+// os.Args if needed
 func run() error {
 	c, err := config.ProvideConfig()
 	if err != nil {
 		return err
 	}
 
-	rules := rule.ProvideRules(c)
+	rules := rule.NewRules(c)
+	// TODO: define any timeouts if you want
 	s := &http.Server{
 		Addr:    ":" + c.ServerPort,
-		Handler: handler.RateLimit(rules)(proxy.TransparentProxy(c.Backend)),
+		Handler: rate.Limit(rules)(proxy.Transparently(c.Backend)),
 	}
 	log.Println("Listening on port: " + c.ServerPort)
 	return s.ListenAndServe()
