+++ pgk/proxy/proxy_test.go
@@ -1,27 +1,21 @@
 package proxy
 
 import (
-	"github.com/dhis2-sre/go-rate-limite/pgk/config"
-	"github.com/stretchr/testify/assert"
 	"net/http"
 	"net/http/httptest"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestTransparentProxyHandler(t *testing.T) {
 	expected := http.StatusOK
 
-	c := &config.Config{
-		Backend: "https://httpbin.org",
-	}
-
-	proxy := ProvideProxy(c)
-
 	req, err := http.NewRequest("GET", "/status/200", nil)
 	assert.NoError(t, err)
 
 	recorder := httptest.NewRecorder()
-	handler := http.HandlerFunc(proxy.TransparentProxyHandler)
+	handler := http.HandlerFunc(Transparently("https://httpbin.org").ServeHTTP)
 
 	handler.ServeHTTP(recorder, req)
