+++ pgk/rule/rule_test.go
@@ -1,9 +1,12 @@
 package rule
 
 import (
+	"net/http"
+	"net/url"
+	"testing"
+
 	"github.com/dhis2-sre/go-rate-limite/pgk/config"
 	"github.com/stretchr/testify/assert"
-	"testing"
 )
 
 func TestPathMatch(t *testing.T) {
@@ -37,4 +40,46 @@ func createRuleWithPathPattern(pathPattern string) *Rule {
 		Handler: nil,
 	}
 	return rule
+}
+
+func TestMatch(t *testing.T) {
+	rule := &config.Rule{
+		PathPattern: "^\\/health$",
+	}
+
+	configRules := []config.Rule{*rule}
+	c := &config.Config{Rules: configRules}
+
+	rules := NewRules(c)
+
+	u, err := url.Parse("http://backend/health")
+	assert.NoError(t, err)
+
+	req := &http.Request{URL: u}
+	actual, _ := rules.Match(req)
+
+	expected := true
+
+	assert.Equal(t, expected, actual)
+}
+
+func TestNoMatch(t *testing.T) {
+	rule := &config.Rule{
+		PathPattern: "^\\/health$",
+	}
+
+	configRules := []config.Rule{*rule}
+	c := &config.Config{Rules: configRules}
+
+	rules := NewRules(c)
+
+	u, err := url.Parse("http://backend/health-no-match")
+	assert.NoError(t, err)
+
+	req := &http.Request{URL: u}
+	actual, _ := rules.Match(req)
+
+	expected := false
+
+	assert.Equal(t, expected, actual)
 }
